Heavily refactored project layout and tree_interpreter.cpp to be more intuitive.
