+++ CMakeLists.txt
@@ -3,10 +3,9 @@ project(PL-JIT-Compiler LANGUAGES CXX)
 
 add_executable(interpreter
         src/main.cpp
-        src/helpers/io_helpers.cpp
-        src/parse_into_ast.cpp
-        src/helpers/debug_helpers.cpp
-        src/tree_interpreter.cpp
+        src/parser/parser.cpp
+        src/ast/ast_printer.cpp
+        src/tree_interpreter/tree_interpreter.cpp
 )
 
 target_include_directories(interpreter PRIVATE third_party)
