--- src/helpers/io_helpers.h
@@ -1,11 +1,0 @@
-#ifndef io_helpers_H
-#define io_helpers_H
-
-#include <filesystem>
-#include <string>
-
-namespace io_helpers {
-  std::string ReadFile(const std::filesystem::path& path);
-}
-
-#endif // io_helpers_H
