--- src/parse_into_ast.h
@@ -1,12 +1,0 @@
-#ifndef parse_into_ast_H
-#define parse_into_ast_H
-
-#include <string>
-
-#include "pl_ast.h"
-
-namespace parse_into_ast {
-  pl_ast::Program ParseFileContentsIntoAST(const std::string& file_contents);
-}
-
-#endif // parse_into_ast_H
