+++ src/tree_interpreter/tree_interpreter.h
@@ -1,7 +1,7 @@
 #ifndef tree_interpreter_H
 #define tree_interpreter_H
 
-#include "pl_ast.h"
+#include "../ast/ast.h"
 
 namespace tree_interpreter {
   void ExecuteAstWithTreeInterpreter(const pl_ast::Program& program);
